QPID-4555: HA Don't shut down on deleting an exchange that is in use as an alternate.
Previously threw an exception in this case which shut down the broker.
Log warning instead, this is not a fatal event.

git-svn-id: https://svn.apache.org/repos/asf/qpid/trunk@1443676 13f79535-47bb-0310-9956-ffa450edef68

+++ qpid/cpp/src/qpid/ha/BrokerReplicator.cpp
@@ -784,6 +784,15 @@ void BrokerReplicator::deleteQueue(const std::string& name, bool purge) {
 
 void BrokerReplicator::deleteExchange(const std::string& name) {
     try {
+        boost::shared_ptr<broker::Exchange> exchange = exchanges.find(name);
+        if (!exchange) {
+            QPID_LOG(warning, logPrefix << "Cannot delete exchange, not found: " << name);
+            return;
+        }
+        if (exchange->inUseAsAlternate()) {
+            QPID_LOG(warning, "Cannot delete exchange, in use as alternate: " << name);
+            return;
+        }
         broker.deleteExchange(name, userId, remoteHost);
         QPID_LOG(debug, logPrefix << "Exchange deleted: " << name);
     } catch (const framing::NotFoundException&) {
